+++ header/DamageCalculator.h
@@ -5,5 +5,5 @@
 
 class DamageCalculator {
     public:
-     static int performCalculation(const Combatant& attacker, const Combatant& defender, const int baseDamage);
+     static double performCalculation(const Combatant& attacker, const Combatant& defender, const double baseDamage);
 };
